File updates

+++ Main.py
@@ -1,32 +1,51 @@
 #!/usr/bin/python
 print "cheacking for updates..."
 import urllib2,os
-response = urllib2.urlopen('https://raw.githubusercontent.com/ZZcat/Update-numbers/master/1')
-html = response.read()
-web_vesion = html[:5]
-f = open('version.txt', 'r')
-vesion = f.read()
-print "Your vesion is ",vesion,"\nThe newest vesion is ", web_vesion
-if vesion == web_vesion:
-   print "You have the newest vesion!"
-else:
-   print "###You need to update this program"
-   print "###Updating..."
-   print "\n###Reading lines"
-   folder_dir = os.popen("pwd").readlines()
-   print "###Done!!!"
-   print "\n###Cuting varuibal"
-   folder_dir = folder_dir[:-10]
-   print "###Done!!!\n\n###Cloning dicrectory"
-   com = "git clone https://github.com/ZZcat/C-code.git"
-   os.system(com)
-   print "###Done\n\n###Moving directory"
-   com = "mv C-code " , folder_dir
-   print "###Done!!!"
-   print "\n###Removing copy folder"
-   com = "rm C-code"
-   print "###Done!!!"
-   
+
+if input("Update(1 or 0):") == True:
+   try:
+      response = urllib2.urlopen('https://raw.githubusercontent.com/ZZcat/Update-numbers/master/1')
+      html = response.read()
+      web_vesion = html[:5]
+      f = open('version.txt', 'r')
+      vesion = f.read()
+      print "Your vesion is ",vesion,"\nThe newest vesion is ", web_vesion
+      if vesion == web_vesion:
+         print "You have the newest vesion!"
+      else:
+         print "###You need to update this program"
+         print "###Updating..."
+         print "\n###Reading lines"
+         folder_dir = os.popen("pwd").readlines()
+         print "###Done!!!"
+         print "\n###Cuting varuibal"
+         folder_dir = folder_dir[:-10]
+         print "###Done!!!\n\n###Cloning dicrectory"
+         com = "git clone https://github.com/ZZcat/C-code.git"
+         os.system(com)
+         print "###Done\n\n###Moving directory"
+         com = "mv C-code " , folder_dir
+         print "###Done!!!"
+         print "\n###Removing copy folder"
+         com = "rm C-code"
+         print "###Done!!!"
+   except:
+         print "###You need to update this program"
+         print "###Updating..."
+         print "\n###Reading lines"
+         folder_dir = os.popen("pwd").readlines()
+         print "###Done!!!"
+         print "\n###Cuting varuibal"
+         folder_dir = folder_dir[:-10]
+         print "###Done!!!\n\n###Cloning dicrectory"
+         com = "git clone https://github.com/ZZcat/C-code.git"
+         os.system(com)
+         print "###Done\n\n###Moving directory"
+         com = "mv C-code " , folder_dir
+         print "###Done!!!"
+         print "\n###Removing copy folder"
+         com = "rm C-code"
+         print "###Done!!!"
 
 
 
@@ -97,7 +116,7 @@ if __name__ == '__main__':
    OS_sys = "win"
    typing = False
    text = ""
-   host = "192.168.253.166"
+   host = raw_input("IP: ")
    port = 9010
 
    # chat vars
@@ -197,7 +216,6 @@ if __name__ == '__main__':
                 text = ""
                 TYPE_TEXT = font.render(text, True, (0, 128, 0))
             elif event.key == pygame.K_1:
-                
                text = text + str(1)
             elif event.key == pygame.K_2:
                text = text + str(2)
@@ -217,6 +235,8 @@ if __name__ == '__main__':
                text = text + str(9)
             elif event.key == (pygame.K_0):
                text = text + str(0)
+            elif event.key == (pygame.K_SPACE):
+               text = text + str(" ")
